♻️ 优化代码，抽离inject环境和脚本执行
